requirements.txt
streamlit>=1.36
torch>=2.2,<3.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
joblib>=1.3
rdkit-pypi>=2022.9.5
